backend/requirements.txt
asgiref==3.3.4
atomicwrites==1.4.0
attrs==21.4.0
backcall==0.2.0
bitstring==3.1.7
blis==0.7.7
catalogue==2.0.7
certifi==2021.5.30
charset-normalizer==2.0.3
click==7.1.2
colorama==0.4.4
coverage==6.3.2
cycler==0.10.0
cymem==2.0.6
debugpy==1.4.3
decorator==5.1.0
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.4
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.11.0
django-crontab==0.7.1
django-extensions==3.1.5
djangorestframework==3.13.1
docker==5.0.0
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.2.0/en_core_web_lg-3.2.0-py3-none-any.whl
entrypoints==0.3
filelock==3.5.0
gunicorn==20.1.0
idna==3.2
iniconfig==1.1.1
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.1.1
jupyter-client==7.0.2
jupyter-core==4.7.1
kiwisolver==1.3.2
langcodes==3.3.0
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
murmurhash==1.0.6
mysql-connector-python==8.0.28
mysqlclient==2.1.0
nest-asyncio==1.5.1
numpy==1.20.2
opencv-python==4.5.5.64
packaging==21.3
pandas==1.2.4
parso==0.8.2
pathy==0.6.1
pickleshare==0.7.5
pikepdf==5.0.1
Pillow==8.3.2
pipenv==2022.1.8
platformdirs==2.5.0
pluggy==1.0.0
preshed==3.0.6
prompt-toolkit==3.0.20
psycopg2==2.9.1
py==1.11.0
pydantic==1.8.2
Pygments==2.10.0
pyparsing==2.4.7
PyQt5==5.15.2
pyqt5-plugins==5.15.2.2.1.0
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
pyqt5-tools==5.15.2.3.0.2
pytesseract==0.3.9
pytest==7.1.1
pytest-cov==3.0.0
python-dateutil==2.8.1
python-dotenv==0.17.0
pytube==12.0.0
pytz==2021.1
pywin32==227
pyzmq==22.2.1
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
requests==2.26.0
selenium==3.141.0
six==1.16.0
smart-open==5.2.1
spacy==3.2.4
spacy-legacy==3.0.9
spacy-loggers==1.0.2
sqlparse==0.4.1
srsly==2.4.3
thinc==8.0.15
tk==0.1.0
tomli==2.0.1
tornado==6.1
tqdm==4.64.0
traitlets==5.1.0
typer==0.4.1
typing_extensions==4.1.1
urllib3==1.25.10
virtualenv==20.13.1
virtualenv-clone==0.5.7
wasabi==0.9.1
wcwidth==0.2.5
websocket-client==1.1.0
whitenoise==6.0.0
XlsxWriter==3.0.3
xlwt==1.3.0
